Copying simplified Keycloak API for mounting volumes instead of supporting all kinds of volume sources
Keycloak Operator supports mounting custom volumes inside of the Keycloak deployment out of either ConfigMaps or Secret, and no other type of source.

Instead of supporting all kinds of volume sources, which enlarges the Authorino Operator API and can lead to higher coupling to K8s core types and more cases to support and maintain, we will follow the same approach adopted by the Keycloak Operator and begin with a simpler API – though probably enough to cover the high majority of cases.