fix(labels): do not overwrite user added labels from Deployment's spec.template
PR #91 brought the ability to propagate labels defined for Authorino CR
down to Deployment.
There was some discussion to have those labels propagated to template and therefore pods which are
which are part of the deployment. However, for now we decided to add this change which adds the ability for a
user added label to not be overwritten by the authorino-operator

Signed-off-by: Laura Fitzgerald <lfitzger@redhat.com>
Co-authored-by: bartoszmajsak <bartosz.majsak@gmail.com>