fix(labels): propagates labels to Deployment's spec.template
PR #91 brought the ability to propagate labels defined for Authorino CR
down to Deployment. However, more often than not, it's more desired to
have those labels propagated to template and therefore pods which are
part of the deployment. This can simplify e.g. making Authorino part of
service mesh, as injection label can be consistently added to relevant
resources.

When creating the deployment, instead of having fixed set of labels for
Deployment's `spec.template` this change appends Authorino CR labels as
well.

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>